+++ zc/src/ZC_gnuplot.c
@@ -15,7 +15,7 @@ char** genGnuplotScript_linespoints(char* dataFileName, char* extension, int fon
 	char** lines = (char**)malloc(24*sizeof(char*));
 	char stringBuffer[ZC_BUFS_LONG];
 	
-	int i = 0;
+	size_t i = 0;
 	for(i=0;i<24;i++)
 	{
 		lines[i] = (char*)malloc(250);
@@ -71,7 +71,7 @@ char** genGnuplotScript_histogram(char* dataFileName, char* extension, int fontS
 	char** lines = (char**)malloc(18*sizeof(char*));
 	char stringBuffer[ZC_BUFS_LONG];
 	
-	int i = 0;
+	size_t i = 0;
 	for(i=0;i<18;i++)
 	{
 		lines[i] = (char*)malloc(250);
@@ -122,7 +122,7 @@ char** genGnuplotScript_lines(char* dataFileName, char* extension, int fontSize,
 	char** lines = (char**)malloc(24*sizeof(char*));
 	char stringBuffer[ZC_BUFS_LONG];
 	
-	int i = 0;
+	size_t i = 0;
 	for(i=0;i<24;i++)
 	{
 		lines[i] = (char*)malloc(250);
@@ -178,7 +178,7 @@ char** genGnuplotScript_fillsteps(char* dataFileName, char* extension, int fontS
 	char** lines = (char**)malloc(19*sizeof(char*));
 	char stringBuffer[ZC_BUFS_LONG];
 	
-	int i = 0;
+	size_t i = 0;
 	for(i=0;i<19;i++)
 	{
 		lines[i] = (char*)malloc(250);
