+++ zc/src/ZC_quicksort.c
@@ -3,9 +3,9 @@
 #include "ZC_quicksort.h"
 #include "zc.h"
 
-int ZC_divide(RateDistElem* list,int start,int end)
+size_t ZC_divide(RateDistElem* list,size_t start,size_t end)
 {
-	int i,j,k,pivot;
+	size_t i,j,k,pivot;
 	RateDistElem temp = NULL; //temp is a pointer
 
 	i = start;
@@ -35,9 +35,9 @@ int ZC_divide(RateDistElem* list,int start,int end)
 	return j;
 }
 
-void ZC_quick_sort(RateDistElem* list,int start,int end)
+void ZC_quick_sort(RateDistElem* list,size_t start,size_t end)
 {
-	int mid;
+	size_t mid;
 
 	if(start < end)
 	{
@@ -47,9 +47,9 @@ void ZC_quick_sort(RateDistElem* list,int start,int end)
 	}
 }
 
-int ZC_divide2(StringElem* list,int start,int end)
+size_t ZC_divide2(StringElem* list,size_t start,size_t end)
 {
-	int i,j,k,pivot;
+	size_t i,j,k,pivot;
 	StringElem temp = NULL; //temp is a pointer
 
 	i = start;
@@ -79,9 +79,9 @@ int ZC_divide2(StringElem* list,int start,int end)
 	return j;
 }
 
-void ZC_quick_sort2(StringElem* list,int start,int end)
+void ZC_quick_sort2(StringElem* list,size_t start,size_t end)
 {
-	int mid;
+	size_t mid;
 
 	if(start < end)
 	{
